+++ android/app/src/main/java/com/anonymous/boop/MainApplication.kt
@@ -16,7 +16,7 @@ import com.facebook.react.defaults.DefaultReactNativeHost
 import expo.modules.ApplicationLifecycleDispatcher
 import expo.modules.ReactNativeHostWrapper
 
-import com.boopble.BleBoopPackage
+import com.boopble.BoopBlePackage
 
 class MainApplication : Application(), ReactApplication {
 
@@ -27,7 +27,7 @@ class MainApplication : Application(), ReactApplication {
             PackageList(this).packages.apply {
               // Packages that cannot be autolinked yet can be added manually here, for example:
               // add(MyReactNativePackage())
-              add(BleBoopPackage())
+              add(BoopBlePackage())
             }
 
           override fun getJSMainModuleName(): String = ".expo/.virtual-metro-entry"
