+++ modules/boop-ble/android/src/main/java/com/boopble/BoopBlePackage.kt
@@ -5,9 +5,9 @@ import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
-class BleBoopPackage : ReactPackage {
+class BoopBlePackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-        return listOf(BleBoopModule(reactContext))
+        return listOf(BoopBleModule(reactContext))
     }
 
     override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
