--- modules/boop-ble/ios/BoopBle-Bridging-Header.h
@@ -1,2 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <React/RCTBridgeModule.h>
