+++ modules/boop-ble/ios/BoopBle.h
@@ -1,5 +1,10 @@
+#ifdef __cplusplus
 #import <BoopBleSpec/BoopBleSpec.h>
+#endif
 
-@interface BoopBle : NSObject <NativeBoopBleSpec>
+@interface BoopBle : NSObject
+#ifdef __cplusplus
+<NativeBoopBleSpec>
+#endif
 
 @end
