+++ modules/boop-ble/ios/BoopBle.podspec
@@ -3,7 +3,7 @@ require "json"
 package = JSON.parse(File.read(File.join(__dir__, "../package.json")))
 
 Pod::Spec.new do |s|
-  s.name         = "react-native-ble-boop"
+  s.name         = "boop-ble"
   s.version      = package["version"]
   s.summary      = package["description"]
   s.homepage     = package["homepage"]
