--- modules/react-native-ble-boop/android/build.gradle
@@ -1,32 +1,0 @@
-apply plugin: "com.android.library"
-apply plugin: "kotlin-android"
-
-def reactNative = rootProject.ext.reactNative
-
-android {
-    compileSdkVersion 34
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 34
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-repositories {
-    google()
-    mavenCentral()
-}
-
-dependencies {
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.8.21"
-    implementation "com.facebook.react:react-native:+"
-}
