--- modules/react-native-ble-boop/lib/commonjs/NativeBleBoopSpec.js
@@ -1,10 +1,0 @@
-"use strict";
-
-Object.defineProperty(exports, "__esModule", {
-  value: true,
-});
-exports.default = void 0;
-var _reactNative = require("react-native");
-var _default = (exports.default =
-  _reactNative.TurboModuleRegistry.getEnforcing("BleBoop"));
-//# sourceMappingURL=NativeBleBoopSpec.js.map
