--- modules/react-native-ble-boop/lib/module/NativeBleBoopSpec.js
@@ -1,3 +1,0 @@
-import { TurboModuleRegistry } from "react-native";
-export default TurboModuleRegistry.getEnforcing("BleBoop");
-//# sourceMappingURL=NativeBleBoopSpec.js.map
