--- modules/react-native-ble-boop/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "react-native-ble-boop",
-  "version": "1.0.0",
-  "files": [
-    "src",
-    "android",
-    "ios",
-    "lib"
-  ],
-  "description": "Native BLE Turbo Module for boop discovery",
-  "main": "src/index.ts",
-  "react-native": "boop/app/index.tsx",
-  "types": "src/index.ts",
-  "scripts": {
-    "build": "tsc",
-    "prepare": "bob build"
-  },
-  "keywords": [
-    "react-native",
-    "ble",
-    "bluetooth",
-    "turbo-module"
-  ],
-  "devDependencies": {
-    "@react-native/codegen": "^0.74.0",
-    "react-native-builder-bob": "^0.23.0",
-    "typescript": "^5.0.0"
-  },
-  "peerDependencies": {
-    "react": "*",
-    "react-native": "*"
-  },
-  "react-native-builder-bob": {
-    "source": "src",
-    "output": "lib",
-    "targets": [
-      "commonjs",
-      "module",
-      [
-        "typescript",
-        {
-          "project": "tsconfig.build.json"
-        }
-      ]
-    ]
-  },
-  "codegenConfig": {
-    "name": "BleBoopSpec",
-    "type": "modules",
-    "jsSrcsDir": "src",
-    "android": {
-      "javaPackageName": "com.boopble"
-    }
-  }
-}
