--- modules/react-native-ble-boop/tsconfig.build.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "declarationMap": true
-  },
-  "exclude": ["**/__tests__/**"]
-}
