--- modules/react-native-ble-boop/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2018",
-    "module": "commonjs",
-    "lib": ["ES2018"],
-    "declaration": true,
-    "outDir": "lib",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "moduleResolution": "node",
-    "baseUrl": ".",
-    "paths": {
-      "react-native-ble-boop": ["modules/react-native-ble-boop/src"]
-    }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "lib"]
-}
