+++ package.json
@@ -31,7 +31,7 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-native": "0.81.4",
-    "react-native-ble-boop": "file:modules/react-native-ble-boop",
+    "boop-ble": "file:modules/boop-ble",
     "react-native-gesture-handler": "~2.28.0",
     "react-native-reanimated": "~4.1.0",
     "react-native-safe-area-context": "~5.6.0",
